+++ accountApp/forms.py
@@ -1,0 +1,6 @@
+from django import forms
+from django.contrib.auth.models import User
+
+class LoginForm(forms.Form):
+    username = forms.CharField()
+    password = forms.CharField(widget=forms.PasswordInput)
