+++ templates/account/log-in.html
@@ -1,0 +1,12 @@
+{% extends "base.html" %}
+{% block title %}Log-in{% endblock %}
+
+{% block content %}
+  <h1>Log-in</h1>
+  <p>Please, use the following form to log-in:</p>
+  <form action="." method="post">
+     {{ form.as_p }}
+     {% csrf_token %}
+     <p><input type="submit" value="Log-in"></p>
+  </form>
+{% endblock %}
