+++ templates/account/profile_view.html
@@ -33,7 +33,7 @@
 
 <div class="page-header row">
 	<div class="col-md-4">
-		<a href="#" role="button" class="btn btn-success">Uplood Transaction</a>
+		<a href="#" role="button" class="btn btn-success">Upload Transaction</a>
 	</div>
 </div>
 </div>
