+++ templates/base.html
@@ -1,16 +1,16 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="description" content="{{meta_description_content}}">
     <meta name="author" content="{{meta_author_content}}">
-		<meta name="keywords" content="{{meta_keywords_content}}">
+    <meta name="keywords" content="{{meta_keywords_content}}">
     <title>{% block title %}{% endblock %}</title>
-		{% load staticfiles %}
+    {% load staticfiles %}
 
-		<!-- Bootstrap Core CSS -->
+        <!-- Bootstrap Core CSS -->
     <link href="{% static "cssjsfonts/css/bootstrap.min.css" %}" rel="stylesheet">
 
     <!-- Custom CSS -->
@@ -18,38 +18,29 @@
 
     <!-- Custom Fonts -->
     <link href="{% static "cssjsfonts/font-awesome/css/font-awesome.min.css" %}" rel="stylesheet" type="text/css">
-		<link href="{% static "cssjsfonts/css/ustdy.css" %}" rel="stylesheet" type="text/css">
-		{% block extra_css %}{% endblock %}
-    {% block extra_head_js %}{% endblock %}
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-        <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-        <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
-    <![endif]-->
+    <link href="{% static "cssjsfonts/css/base.css" %}" rel="stylesheet" type="text/css">
 </head>
-<body>
-<div id="wrap">
-  {% include 'nav.html' %}
 
-  {% if show_carousel %}
-  <section id="index-header" class="section-idx">
-    	<div class="container-fluid">
-  		  <div class="row">
-  		    <div class="col-lg-12">
-  		        <h1 class="animation animated-item-1 text-center ustdy-header-three"><strong>Wasisi</strong></h1>
-  				    <h3 class="animation animated-item-2 text-center section-subheading">Science & Engineering University Courses, Tutors & More</h3>
-  		    </div>
-  		</div>
+<body>
+    <div id="header">
+    {% include 'nav.html' %}
+    
+    {% if messages %}
+        <ul class="messages">
+            {% for message in messages %}
+                <li class="{{ message.tags }}">
+                    {{ message|safe }}
+                    <a href="#" class="close">×</a>
+                </li>
+            {% endfor %}
+        </ul>
+    {% endif %}
+
+ <!-- Page Content -->    
+    <div id="content">
+        {% block content %}
+        {% endblock %}
     </div>
-  </section>
-  {% endif %}
-  <!-- Page Content -->
-  {% block content %}
-  {% endblock %}
-  <div id="push"></div>
-</div>
 {% include 'footer.html' %}
 </body>
 </html>
