+++ templates/nav.html
@@ -1,34 +1,34 @@
-<nav class="navbar navbar-default navbar-fixed-top" role="navigation">
-  <div class="container">
-    <div class="navbar-header"><!-- Brand and toggle get grouped for better mobile display -->
+<nav class="navbar" role="navigation">
+    <div class="container">
+        <div class="navbar-header"><!-- Brand and toggle get grouped for better mobile display -->
        <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
           <span class="sr-only">Toggle navigation</span>
           <span class="icon-bar"></span>
           <span class="icon-bar"></span>
           <span class="icon-bar"></span>
         </button>
-				<a class="navbar-brand" href="#"><strong  id="ustdy-logo" class="ustdy-header-three">Wasisi</strong></a>
+        <a class="navbar-brand" href="#"><strong  id="ustdy-logo"><span class="logo">Wasisi</strong></span></a>
       </div>
-            <!-- Collect the nav links, forms, and other content for toggling -->
-      <div class="collapse navbar-collapse pull-right" id="bs-example-navbar-collapse-1">
-        <ul class="nav navbar-nav navbar-left">
-          <li {% if selectcourses %}class="active" {%endif%}>
-						<a href="#" >Coffee</a>
-					</li>
-          <li {% if selectlibrary %}class="active" {%endif%}>
-		         <a href="#">Coffee Analytics</a>
-		      </li>
-          <li {% if selecttutoring %}class="active" {%endif%}>
-					  <a href="#">Contact</a>
-					</li>
-          <li {%if selectcontact == True %} class="active" {% endif %}>
-            <a href="#" >Developer</a>
-          </li>
-
-          <li>
-            <a href="{% url 'accountApp:log_in' %}" >Log in</a>
-          </li>
+        {% if request.user.is_authenticated %} <!-- /.display site menu only to authenticated users -->
+        <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+        <ul class="menu">
+          <!-- We check the current section to add a selected class attribute to the corresponding <li> item in order to highlight the current section in the menu using CSS. -->
+            <li {% if section == "dashboard" %}class="selected"{% endif %}><a href="{% url "dashboard" %}">My dashboard</a></li>
+            <li {% if section == "images" %}class="selected"{% endif %}><a href="#">Images</a></li>
+            <li {% if section == "people" %}class="selected"{% endif %}><a href="#">People</a></li>
+            <li {% if section == "coffee analytics" %}class="selected"{% endif %}><a href="#">Coffee Analytics</a></li>
+            <li {% if section == "contacts" %}class="selected" {% endif %}><a href="#">Contact</a></li>
         </ul>
-      </div><!-- /.navbar-collapse -->
-  </div><!-- /.container -->
+        {% endif %}
+        
+        <span class="user">
+        {% if request.user.is_authenticated %}
+            Hello {{ request.user.first_name }}, <a href="{% url "logout" %}">Logout</a>
+        {% else %}
+            <a href="{% url "login" %}">Log-in</a>
+        {% endif %}
+        </span>
+        </div><!-- /.navbar-collapse -->
+    </div> <!-- /.container -->
+    </div> <!-- /.header -->
 </nav>
