+++ templates/nav_archive.html
@@ -1,0 +1,34 @@
+<nav class="navbar navbar-default navbar-fixed-top" role="navigation">
+  <div class="container">
+    <div class="navbar-header"><!-- Brand and toggle get grouped for better mobile display -->
+       <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
+          <span class="sr-only">Toggle navigation</span>
+          <span class="icon-bar"></span>
+          <span class="icon-bar"></span>
+          <span class="icon-bar"></span>
+        </button>
+                <a class="navbar-brand" href="#"><strong  id="ustdy-logo" class="ustdy-header-three">Wasisi</strong></a>
+      </div>
+            <!-- Collect the nav links, forms, and other content for toggling -->
+      <div class="collapse navbar-collapse pull-right" id="bs-example-navbar-collapse-1">
+        <ul class="nav navbar-nav navbar-left">
+          <li {% if selectcourses %}class="active" {%endif%}>
+                        <a href="#" >Coffee</a>
+                    </li>
+          <li {% if selectlibrary %}class="active" {%endif%}>
+                 <a href="#">Coffee Analytics</a>
+              </li>
+          <li {% if selecttutoring %}class="active" {%endif%}>
+                      <a href="#">Contact</a>
+                    </li>
+          <li {%if selectcontact == True %} class="active" {% endif %}>
+            <a href="#" >Developer</a>
+          </li>
+
+          <li>
+            <a href="{% url 'login' %}" >Log in</a>
+          </li>
+        </ul>
+      </div><!-- /.navbar-collapse -->
+  </div><!-- /.container -->
+</nav>
