+++ templates/registration/logged_out.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block title %}Logged out{% endblock %}
+
+{% block content %}
+    <h1>Logged out</h1>
+    <p>You have been successfully logged out. You can <a href="{% url "login" %}">log-in again</a>.</p>
+{% endblock %}
