+++ templates/registration/password_change_done.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+{% block title %}Password changed{% endblock %}
+{% block content %}
+<h1 class="page-header text-center">Password changed</h1>
+<p class="text-center">Your password has been successfully changed.</p>
+{% endblock %}
