+++ templates/registration/password_change_form.html
@@ -1,0 +1,13 @@
+{% extends "base.html" %}
+
+{% block title %}Change your password{% endblock %}
+
+{% block content %}
+    <h1>Change you password</h1>
+    <p>Use the form below to change your password.</p>
+    <form action="." method="post">
+        {{ form.as_p }}
+        {% csrf_token %}
+        <button type="submit" class="btn btn-primary" >Change</button>
+    </form>
+{% endblock %}
