+++ templates/registration/password_reset_complete.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block title %}Password reset{% endblock %}
+
+{% block content %}
+    <h1>Password set</h1>
+    <p>Your password has been set. You can <a href="{% url "login" %}">log in now</a></p>
+{% endblock %}
