+++ templates/registration/password_reset_confirm.html
@@ -1,0 +1,19 @@
+{% extends "base.html" %}
+
+{% block title %}Reset your password{% endblock %}
+
+{% block content %}
+    <h1>Reset your password</h1>
+    {% if validlink %}
+        <p>Please enter your new password twice:</p>
+        <form action="." method="post">
+            {{ form.as_p }}
+            {% csrf_token %}
+           <!-- <p><input type="submit" value="Change my password" /></p> -->
+
+            <button type="submit" class="btn btn-primary" >Change my password</button>
+        </form>
+    {% else %}
+        <p>The password reset link was invalid, possibly because it has already been used. Please request a new password reset.</p>
+    {% endif %}
+{% endblock %}
