+++ templates/registration/password_reset_done.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block title %}Reset your password{% endblock %}
+
+{% block content %}
+    <h1 class="page-header">Reset your password</h1>
+    <p>We've emailed you instructions for setting your password.</p>
+    <p>If you don't receive an email, please make sure you've entered the address you registered with.</p>
+{% endblock %}
