+++ templates/registration/password_reset_form.html
@@ -1,0 +1,30 @@
+{% extends "base.html" %}
+
+{% block title %}Reset your password{% endblock %}
+
+{% block content %}
+
+<!--<h1 class="page-header">Forgot your password?</h1>-->
+ 
+<div class="row page-header">
+  <div class="col-md-6 col-md-offset-2 text-center">    
+    <h1>Forgotten your password?</h1>
+    </div>
+</div>
+
+<div class="col-md-6 col-md-offset-2">
+<div class="text-center">
+    <p class="help-block">Enter your e-mail address to obtain a new password.</p>
+    <p class="help-block">We will send you an email with a link to reset your password.</p>
+</div>
+</div>
+
+<div class="col-md-6 col-md-offset-2">
+    <form action="." method="post">
+        {{ form.as_p }}
+       <!-- <p><input type="submit" value="Send e-mail"></p> -->
+        <button type="submit" class="btn btn-primary" >Send e-mail</button> <button type="submit" class="btn btn-primary" name="cancel" formnovalidate>Cancel</button>
+        {% csrf_token %}
+    </form>
+</div>    
+{% endblock %}
